+++ NEWS.md
@@ -1,4 +1,5 @@
 # Version 1.0.9000.9061
+* Renamed slots in `NextBestTDsamples` class.
 * Removed `Model` class.
 * Renamed `SIM` argument of `nextBest` methods to `in_sim`.
 * Renamed the elements of the list returned by `nextBest` methods for `NextBestMaxGain`